+++ index.js
@@ -2,8 +2,9 @@
 
 var eslint = require("eslint").linter
 var Config = require("eslint/lib/config");
-// var loaderUtils = require("loader-utils")
-//
+
+var stylish = require("eslint/lib/formatters/stylish")
+
 /**
  * linter
  *
@@ -12,27 +13,28 @@ var Config = require("eslint/lib/config");
  * @param  {Object} webpack
  */
 function lint(input, config, webpack) {
-  var messages = []
+  var res = eslint.verify(input, config)
+  if (res.length) {
+    var reporter = webpack.options.eslint.reporter || function(results) {
+      // in order to use eslint formatters
+      // we need to reproduce the object passed to them
+      var msgs = stylish([{
+        filePath: "",
+        messages: results
+      }]).split("\n")
+      // drop the line that should contains filepath we do not have
+      msgs.splice(0, 1)
 
-  eslint.verify(input, config).forEach(function eslintMessageCallback(m) {
-    messages.push(
-      "  " + m.message +
-      "(" + m.ruleId + ")" +
-      " @ line " + m.line + " column " + m.column +
-      " - " +
-      (m.fatal ? "fatal, " : "") +
-      "severity: " + m.severity
-    )
-  })
-
-  if (messages.length) {
-    var message = messages.join("\n\n")
+      return msgs.join("\n")
+    }
+    var messages = reporter(res)
     var emitter = webpack.options.eslint.emitErrors ? webpack.emitError : webpack.emitWarning
+
     if (emitter) {
-      emitter("eslint results in errors\n" + message)
+      emitter(messages)
     }
     else {
-      throw new Error("Your module system doesn't support emitWarning. Update available? \n" + message)
+      throw new Error("Your module system doesn't support emitWarning. Update available? \n" + messages)
     }
   }
 }